+++ lib/src/ast/sass/statement/include_rule.dart
@@ -6,7 +6,7 @@ import 'package:source_span/source_span.dart';
 
 import '../../../util/span.dart';
 import '../../../visitor/interface/statement.dart';
-import '../argument_invocation.dart';
+import '../argument_list.dart';
 import '../callable_invocation.dart';
 import '../reference.dart';
 import '../statement.dart';
@@ -30,7 +30,7 @@ final class IncludeRule extends Statement
   final String originalName;
 
   /// The arguments to pass to the mixin.
-  final ArgumentInvocation arguments;
+  final ArgumentList arguments;
 
   /// The block that will be invoked for [ContentRule]s in the mixin being
   /// invoked, or `null` if this doesn't pass a content block.