+++ pkg/sass-parser/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.8-dev
+## 0.4.8
 
 Add support for parsing the `@mixin` rule.